Implemented AMkd_encode
Also made some small fixes and updated docs
