+++ docs/globals_vars.html
@@ -23,7 +23,7 @@
   <td id="projectlogo"><img alt="Logo" src="AMkd_group.ico"/></td>
   <td id="projectalign" style="padding-left: 0.5em;">
    <div id="projectname">AMkd
-   &#160;<span id="projectnumber">0.4.0</span>
+   &#160;<span id="projectnumber">0.5.0</span>
    </div>
    <div id="projectbrief">CNV codec</div>
   </td>
@@ -66,10 +66,10 @@ $(function() {
 <div class="contents">
 &#160;<ul>
 <li>AMKD_CONFIG_DEFAULT
-: <a class="el" href="AMkd_8h.html#aa9f3a2f498364667204262dbaba7b66b">AMkd.h</a>
+: <a class="el" href="AMkd_8h.html#a5330ab68322cf515ad63f88f8112b8d2">AMkd.h</a>
 </li>
 <li>AMKD_CONFIG_NONE
-: <a class="el" href="AMkd_8h.html#ad763a6992dd287abd8ab94710d8b5b1c">AMkd.h</a>
+: <a class="el" href="AMkd_8h.html#ac734f266be9849d0654bdbe22a472665">AMkd.h</a>
 </li>
 </ul>
 </div><!-- contents -->
