+++ docs/search/all_0.js
@@ -2,8 +2,8 @@ var searchData=
 [
   ['amkd_2eh_0',['AMkd.h',['../AMkd_8h.html',1,'']]],
   ['amkd_5fconfig_1',['AMkd_config',['../structAMkd__config.html',1,'AMkd_config'],['../AMkd_8h.html#a5d82d4841948c96d2aeb22405d42876d',1,'AMkd_config():&#160;AMkd.h']]],
-  ['amkd_5fconfig_5fdefault_2',['AMKD_CONFIG_DEFAULT',['../AMkd_8h.html#aa9f3a2f498364667204262dbaba7b66b',1,'AMkd.h']]],
-  ['amkd_5fconfig_5fnone_3',['AMKD_CONFIG_NONE',['../AMkd_8h.html#ad763a6992dd287abd8ab94710d8b5b1c',1,'AMkd.h']]],
+  ['amkd_5fconfig_5fdefault_2',['AMKD_CONFIG_DEFAULT',['../AMkd_8h.html#a5330ab68322cf515ad63f88f8112b8d2',1,'AMkd.h']]],
+  ['amkd_5fconfig_5fnone_3',['AMKD_CONFIG_NONE',['../AMkd_8h.html#ac734f266be9849d0654bdbe22a472665',1,'AMkd.h']]],
   ['amkd_5fdecode_4',['AMkd_decode',['../AMkd_8h.html#ac9e9634d274575fce942d507a755feb2',1,'AMkd.c']]],
   ['amkd_5fdetect_5fencoding_5',['AMkd_detect_encoding',['../AMkd_8h.html#a33db0504636c3c71f93f4192d441d246',1,'AMkd.c']]],
   ['amkd_5fencode_6',['AMkd_encode',['../AMkd_8h.html#aa2300276c1f678c206831d1e37013f2f',1,'AMkd.c']]],
