+++ docs/search/variables_0.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['amkd_5fconfig_5fdefault_24',['AMKD_CONFIG_DEFAULT',['../AMkd_8h.html#aa9f3a2f498364667204262dbaba7b66b',1,'AMkd.h']]],
-  ['amkd_5fconfig_5fnone_25',['AMKD_CONFIG_NONE',['../AMkd_8h.html#ad763a6992dd287abd8ab94710d8b5b1c',1,'AMkd.h']]]
+  ['amkd_5fconfig_5fdefault_24',['AMKD_CONFIG_DEFAULT',['../AMkd_8h.html#a5330ab68322cf515ad63f88f8112b8d2',1,'AMkd.h']]],
+  ['amkd_5fconfig_5fnone_25',['AMKD_CONFIG_NONE',['../AMkd_8h.html#ac734f266be9849d0654bdbe22a472665',1,'AMkd.h']]]
 ];
