+++ src/amkd.c
@@ -4,6 +4,9 @@
 #include <string.h>
 #include "../include/AMkd.h"
 
+const AMkd_config AMKD_CONFIG_DEFAULT = {6, 'C'};
+const AMkd_config AMKD_CONFIG_NONE = {0, 'C'};
+
 typedef enum AMkd_config_compare_mode {
     AMKD_COMPARE_WHOLE,
     AMKD_COMPARE_STEPS,
@@ -77,7 +80,7 @@ AMkd_error AMkd_decode(const char *encoded_str, char **decoded_str, AMkd_config 
         step = 1,
         decoded_str_max_size = AMkd_calculate_size(encoded_str, settings, AMKD_CONFIG_NONE);
     *decoded_str = (char *)malloc((decoded_str_max_size) * sizeof(char));
-    if (decoded_str == NULL) {
+    if (*decoded_str == NULL) {
         return AMKD_ERROR_OUT_OF_MEMORY;
     }
     {
@@ -105,7 +108,7 @@ AMkd_error AMkd_decode(const char *encoded_str, char **decoded_str, AMkd_config 
                 }
                 if (sequence_completed) {
                     (*decoded_str)[decoded_index] = '\n';
-                    encoded_index++;
+                    encoded_index += 2;
                 } else {
                     (*decoded_str)[decoded_index] = encoded_str[encoded_index] + movement;
                     step++;
@@ -115,11 +118,19 @@ AMkd_error AMkd_decode(const char *encoded_str, char **decoded_str, AMkd_config 
                 step++;
             }
             encoded_index++;
+            if ((*decoded_str)[decoded_index] < 32) {
+                local_errno = AMKD_WARNING_CONTROL_CHARS;
+            }
             decoded_index++;
             if (step > settings.step_count) {
                 step = 1;
             }
         }
+    }
+    if (decoded_index < decoded_str_max_size) {
+        (*decoded_str)[decoded_index] = '\0';
+    } else {
+        (*decoded_str)[decoded_str_max_size] = '\0';
     }
     return local_errno;
 }
@@ -129,9 +140,51 @@ AMkd_error AMkd_encode(const char *decoded_str, char **encoded_str, AMkd_config 
     if (AMkd_compare_configs(settings, AMKD_CONFIG_NONE, AMKD_COMPARE_STEPS)) {
         return AMKD_ERROR_MISSING_SETTINGS;
     }
-    /*
-        [TODO]
-    */
+    unsigned decoded_length = strlen(decoded_str),
+        decoded_index = 0,
+        encoded_index = 0,
+        step = 1,
+        line_break_count = 0,
+        encoded_str_max_size = AMkd_calculate_size(decoded_str, AMKD_CONFIG_NONE, settings);
+    *encoded_str = (char *)malloc((encoded_str_max_size) * sizeof(char));
+    if (*encoded_str == NULL) {
+        return AMKD_ERROR_OUT_OF_MEMORY;
+    }
+    while (decoded_index < decoded_length && encoded_index < encoded_str_max_size) {
+        int movement = (step + 1) / 2;
+        if (step % 2 == 0) {
+            movement *= -1;
+        }
+        if (decoded_str[decoded_index] == '\n') {
+            if (encoded_index + 2 < encoded_str_max_size) {
+                (*encoded_str)[encoded_index] = '<';
+                (*encoded_str)[encoded_index + 1] = 'E';
+                (*encoded_str)[encoded_index + 2] = '>';
+            }
+            encoded_index += 2;
+            line_break_count++;
+            //makes literal line break every [step_count] coded line breaks
+            if (line_break_count >= settings.step_count) {
+                if (encoded_index + 1 < encoded_str_max_size) {
+                    (*encoded_str)[encoded_index] = '\n';
+                    encoded_index++;
+                }
+            }
+        } else {
+            (*encoded_str)[encoded_index] = decoded_str[decoded_index] + movement;
+            step++;
+            if (step > settings.step_count) {
+                step = 1;
+            }
+        }
+        encoded_index++;
+        decoded_index++;
+    }
+    if (encoded_index < encoded_str_max_size) {
+        (*encoded_str)[encoded_index] = '\0';
+    } else {
+        (*encoded_str)[encoded_str_max_size] = '\0';
+    }
     return AMKD_ERROR_NONE;
 }
